+++ Sources/RichTextKit/Commands/RichTextCommand.swift
@@ -1,0 +1,19 @@
+//
+//  RichTextCommands.swift
+//  RichTextKit
+//
+//  Created by Daniel Saidi on 2024-01-18.
+//  Copyright © 2024 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This type is used as a namespace for command-specific types,
+ like views that are to be added to the main menu bar.
+ 
+ > Important: Most rich text commands require that there's a
+ focused ``RichTextContext`` value in the view hierarchy. If
+ not, they will be disabled.
+ */
+public struct RichTextCommand {}